+++ src/Profile/Shopware6/Converter/PageSystemConfigConverter.php
@@ -10,6 +10,8 @@ namespace SwagMigrationAssistant\Profile\Shopware6\Converter;
 use Shopware\Core\Framework\Log\Package;
 use SwagMigrationAssistant\Migration\Converter\ConvertStruct;
 use SwagMigrationAssistant\Migration\DataSelection\DefaultEntities;
+use SwagMigrationAssistant\Migration\Logging\Log\Builder\MigrationLogBuilder;
+use SwagMigrationAssistant\Migration\Logging\Log\ConvertAssociationMissingLog;
 use SwagMigrationAssistant\Migration\Logging\LoggingServiceInterface;
 use SwagMigrationAssistant\Migration\Mapping\Lookup\SystemConfigLookup;
 use SwagMigrationAssistant\Migration\Mapping\MappingServiceInterface;
@@ -46,6 +48,14 @@ class PageSystemConfigConverter extends ShopwareConverter
             );
 
             if ($mappedSalesChannelId === null) {
+                $this->loggingService->log(
+                    MigrationLogBuilder::fromMigrationContext($this->migrationContext)
+                        ->withEntityName(DefaultEntities::PAGE_SYSTEM_CONFIG)
+                        ->withFieldName('salesChannelId')
+                        ->withSourceData($data)
+                        ->build(ConvertAssociationMissingLog::class)
+                );
+
                 return new ConvertStruct(null, $data);
             }
         }