+++ tests/Profile/Shopware6/Converter/SalesChannelConverterTest.php
@@ -13,6 +13,7 @@ use Shopware\Core\Framework\Log\Package;
 use SwagMigrationAssistant\Migration\Converter\ConverterInterface;
 use SwagMigrationAssistant\Migration\Converter\ConvertStruct;
 use SwagMigrationAssistant\Migration\DataSelection\DataSet\DataSet;
+use SwagMigrationAssistant\Migration\DataSelection\DefaultEntities;
 use SwagMigrationAssistant\Migration\Logging\LoggingServiceInterface;
 use SwagMigrationAssistant\Migration\Mapping\Lookup\SalesChannelLookup;
 use SwagMigrationAssistant\Migration\Mapping\Lookup\SalesChannelTypeLookup;
@@ -74,7 +75,7 @@ class SalesChannelConverterTest extends ShopwareConverterTest
         unset($translation);
 
         $salesChannelTypeLookup = $this->createMock(SalesChannelTypeLookup::class);
-        $salesChannelTypeLookup->method('get')->willReturn($input['typeId']);
+        $salesChannelTypeLookup->method('hasType')->willReturn(true);
 
         $salesChannelLookup = $this->createMock(SalesChannelLookup::class);
         $salesChannelLookup->method('getSalesChannelWithTypeAndName')
@@ -101,6 +102,92 @@ class SalesChannelConverterTest extends ShopwareConverterTest
         static::assertSame('Storefront (Migration)', $output['translations'][0]['name']);
     }
 
+    public function testConvertMapsExistingCustomTypeSalesChannel(): void
+    {
+        $input = require __DIR__ . '/../../../_fixtures/Shopware6/SalesChannel/04-UnsupportedType/input.php';
+        $input['typeId'] = '9ce0868f406d47d98cfe4b281e62f098';
+
+        $existingSalesChannelId = '11111111111111111111111111111111';
+
+        $salesChannelTypeLookup = $this->createMock(SalesChannelTypeLookup::class);
+        $salesChannelTypeLookup->method('hasType')->with($input['typeId'], static::anything())->willReturn(true);
+
+        $salesChannelLookup = $this->createMock(SalesChannelLookup::class);
+        $salesChannelLookup->method('getSalesChannelWithTypeAndName')
+            ->with($input['typeId'], $input['name'], static::anything())
+            ->willReturn($existingSalesChannelId);
+
+        $this->converter = new SalesChannelConverter(
+            $this->mappingService,
+            $this->loggingService,
+            $salesChannelTypeLookup,
+            $salesChannelLookup
+        );
+
+        $context = Context::createDefaultContext();
+        $convertResult = $this->converter->convert($input, $context, $this->migrationContext);
+
+        static::assertInstanceOf(ConvertStruct::class, $convertResult);
+        static::assertNull($convertResult->getConverted());
+
+        $mapping = $this->mappingService->getMapping(
+            $this->migrationContext->getConnection()->getId(),
+            SalesChannelDataSet::getEntity(),
+            $input['id'],
+            $context
+        );
+
+        static::assertNotNull($mapping);
+        static::assertSame($existingSalesChannelId, $mapping['entityId']);
+        static::assertSame([], $this->loggingService->getLoggingArray());
+    }
+
+    public function testConvertCreatesNewSalesChannelForExistingCustomTypeWithoutMatchingTargetChannel(): void
+    {
+        $input = require __DIR__ . '/../../../_fixtures/Shopware6/SalesChannel/01-HappyCase/input.php';
+        $mappingArray = require __DIR__ . '/../../../_fixtures/Shopware6/SalesChannel/03-DefaultSalesChannelWithMapping/mapping.php';
+        $expectedOutput = require __DIR__ . '/../../../_fixtures/Shopware6/SalesChannel/01-HappyCase/output.php';
+
+        $input['typeId'] = '9ce0868f406d47d98cfe4b281e62f098';
+        $expectedOutput['typeId'] = '9ce0868f406d47d98cfe4b281e62f098';
+
+        $salesChannelTypeLookup = $this->createMock(SalesChannelTypeLookup::class);
+        $salesChannelTypeLookup->method('hasType')->with($input['typeId'], static::anything())->willReturn(true);
+
+        $salesChannelLookup = $this->createMock(SalesChannelLookup::class);
+        $salesChannelLookup->method('getSalesChannelWithTypeAndName')
+            ->with($input['typeId'], $input['name'], static::anything())
+            ->willReturn(null);
+
+        $this->converter = new SalesChannelConverter(
+            $this->mappingService,
+            $this->loggingService,
+            $salesChannelTypeLookup,
+            $salesChannelLookup
+        );
+
+        $this->loadMapping($mappingArray);
+
+        $context = Context::createDefaultContext();
+        $convertResult = $this->converter->convert($input, $context, $this->migrationContext);
+
+        static::assertInstanceOf(ConvertStruct::class, $convertResult);
+
+        $output = $convertResult->getConverted();
+        static::assertSame($expectedOutput, $output);
+
+        $mapping = $this->mappingService->getMapping(
+            $this->migrationContext->getConnection()->getId(),
+            DefaultEntities::SALES_CHANNEL,
+            $input['id'],
+            $context
+        );
+
+        static::assertNotNull($mapping);
+        static::assertSame($input['id'], $mapping['entityId']);
+        static::assertSame([], $this->loggingService->getLoggingArray());
+    }
+
     protected function createConverter(
         MappingServiceInterface $mappingService,
         LoggingServiceInterface $loggingService,
@@ -108,13 +195,13 @@ class SalesChannelConverterTest extends ShopwareConverterTest
         ?array $mappingArray = [],
     ): ConverterInterface {
         $salesChannelTypeLookup = $this->createMock(SalesChannelTypeLookup::class);
-        $salesChannelTypeLookup->method('get')->willReturnCallback(
-            static function (string $salesChannelTypeId, mixed $context): ?string {
+        $salesChannelTypeLookup->method('hasType')->willReturnCallback(
+            static function (string $salesChannelTypeId, mixed $context): bool {
                 if ($salesChannelTypeId === 'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa') {
-                    return null;
+                    return false;
                 }
 
-                return $salesChannelTypeId;
+                return true;
             }
         );